+++ packages/delta-protocol-repository-ws/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.8.0 — not yet released
+
+* Fix `messageKind` field for request messages of queries to *not* have the "`Request`" suffix.
+
+
+## 0.7.2
+
+(No changes)
+
+
 ## 0.7.1
 
 (The 0.7.0 release was deprecated because its `validation` package was faulty.)